+++ tests/trainer/test_base_trainer.py
@@ -6,8 +6,9 @@ import tensorflow as tf
 from mock import Mock
 
 from tfsnippet.dataflows import DataFlow
-from tfsnippet.scaffold import TrainLoop, AnnealingVariable
+from tfsnippet.scaffold import TrainLoop, AnnealingVariable, EventKeys
 from tfsnippet.trainer import *
+from tfsnippet.utils import EventSource
 
 
 class BaseTrainerTestCase(tf.test.TestCase):
@@ -16,12 +17,7 @@ class BaseTrainerTestCase(tf.test.TestCase):
         loop = Mock(valid_metric_name='valid_loss')
         t = BaseTrainer(loop)
         self.assertIs(loop, t.loop)
-        self.assertEqual(
-            (t.before_epochs, t.before_steps, t.after_steps, t.after_epochs),
-            t.hook_lists
-        )
-        for hkl in t.hook_lists:
-            self.assertIsInstance(hkl, HookList)
+        self.assertIsInstance(t.events, EventSource)
 
     def test_add_and_remove_hooks(self):
         loop = Mock(
@@ -69,40 +65,50 @@ class BaseTrainerTestCase(tf.test.TestCase):
             epochs=18
         )
 
-        self.assertEqual('HookList()', repr(t.before_steps))
-        self.assertEqual('HookList()', repr(t.before_epochs))
-        steps_ans = 'HookList(eval_step:5,{!r}:9,eval_step2:15,anneal_step:7,' \
-                    '{!r}:11,anneal_step2:17,print_logs:3,print_logs:13)'. \
-                    format(eval1.run, anneal1.anneal)
-        self.assertEqual(steps_ans, repr(t.after_steps))
-        epochs_ans = 'HookList(eval_epoch:6,{!r}:10,eval_epoch2:16,' \
-                     'anneal_epoch:8,{!r}:12,anneal_epoch2:18,print_logs:4,' \
-                     'print_logs:14)'. \
-                     format(eval2.run, anneal2.anneal)
-        self.assertEqual(epochs_ans, repr(t.after_epochs))
+        self.assertEqual(
+            repr(t.events._event_handlers_map[EventKeys.AFTER_STEP_EVAL]),
+            '[eval_step:5, {!r}:9, eval_step2:15]'.format(eval1.run)
+        )
+        self.assertEqual(
+            repr(t.events._event_handlers_map[EventKeys.AFTER_STEP_ANNEAL]),
+            '[anneal_step:7, {!r}:11, anneal_step2:17]'.format(anneal1.anneal)
+        )
+        self.assertEqual(
+            repr(t.events._event_handlers_map[EventKeys.AFTER_STEP_LOG]),
+            '[print_logs:3, print_logs:13]'
+        )
+
+        self.assertEqual(
+            repr(t.events._event_handlers_map[EventKeys.AFTER_EPOCH_EVAL]),
+            '[eval_epoch:6, {!r}:10, eval_epoch2:16]'.format(eval2.run)
+        )
+        self.assertEqual(
+            repr(t.events._event_handlers_map[EventKeys.AFTER_EPOCH_ANNEAL]),
+            '[anneal_epoch:8, {!r}:12, anneal_epoch2:18]'.format(anneal2.anneal)
+        )
+        self.assertEqual(
+            repr(t.events._event_handlers_map[EventKeys.AFTER_EPOCH_LOG]),
+            '[print_logs:4, print_logs:14]'
+        )
 
         # test remove
         t.remove_log_hooks()
-        steps_ans = 'HookList(eval_step:5,{!r}:9,eval_step2:15,anneal_step:7,' \
-                    '{!r}:11,anneal_step2:17)'. \
-                    format(eval1.run, anneal1.anneal)
-        self.assertEqual(steps_ans, repr(t.after_steps))
-        epochs_ans = 'HookList(eval_epoch:6,{!r}:10,eval_epoch2:16,' \
-                     'anneal_epoch:8,{!r}:12,anneal_epoch2:18)'. \
-                     format(eval2.run, anneal2.anneal)
-        self.assertEqual(epochs_ans, repr(t.after_epochs))
+        self.assertNotIn(
+            EventKeys.AFTER_STEP_LOG, t.events._event_handlers_map)
+        self.assertNotIn(
+            EventKeys.AFTER_EPOCH_LOG, t.events._event_handlers_map)
 
         t.remove_validation_hooks()
-        steps_ans = 'HookList(anneal_step:7,{!r}:11,anneal_step2:17)'. \
-                    format(anneal1.anneal)
-        self.assertEqual(steps_ans, repr(t.after_steps))
-        epochs_ans = 'HookList(anneal_epoch:8,{!r}:12,anneal_epoch2:18)'. \
-                     format(anneal2.anneal)
-        self.assertEqual(epochs_ans, repr(t.after_epochs))
+        self.assertNotIn(
+            EventKeys.AFTER_STEP_EVAL, t.events._event_handlers_map)
+        self.assertNotIn(
+            EventKeys.AFTER_EPOCH_EVAL, t.events._event_handlers_map)
 
         t.remove_annealing_hooks()
-        self.assertEqual('HookList()', repr(t.after_steps))
-        self.assertEqual('HookList()', repr(t.after_epochs))
+        self.assertNotIn(
+            EventKeys.AFTER_STEP_ANNEAL, t.events._event_handlers_map)
+        self.assertNotIn(
+            EventKeys.AFTER_EPOCH_ANNEAL, t.events._event_handlers_map)
 
         # test error add
         func_list = [
@@ -121,27 +127,46 @@ class BaseTrainerTestCase(tf.test.TestCase):
                                           '`steps` should be specified'):
                     func(**kwargs)
 
+    def test_hook_freq(self):
+        loop = Mock(
+            valid_metric_name='valid_loss',
+            print_logs=Mock(return_value=None, __repr__=lambda o: 'print_logs')
+        )
+        t = BaseTrainer(loop)
+        f = Mock()
+        t.evaluate_after(f, steps=5)
+
+        for i in range(1, 6):
+            t.events.fire(EventKeys.AFTER_STEP_EVAL, i)
+        t.events.fire(EventKeys.AFTER_STEP_EVAL, 7)
+        t.events.fire(EventKeys.AFTER_STEP_EVAL, 10)
+
+        self.assertEqual(f.call_count, 2)
+
     def test_run(self):
         with self.test_session() as session:
             df = DataFlow.arrays([np.arange(6, dtype=np.float32)], batch_size=4)
 
-            def log_message(m):
-                logged_messages.append(m)
-            logged_messages = []
+            def log_event(m, *args):
+                logged_events.append((m,) + args)
+            logged_events = []
 
             # test default loss weight and merged feed dict
             with TrainLoop([], max_epoch=2) as loop:
                 t = BaseTrainer(loop)
                 t._run_step = Mock(return_value=None)
                 t._iter_steps = Mock(wraps=lambda: loop.iter_steps(df))
-                t.before_epochs.add_hook(
-                    functools.partial(log_message, 'before_epoch'))
-                t.before_steps.add_hook(
-                    functools.partial(log_message, 'before_step'))
-                t.after_steps.add_hook(
-                    functools.partial(log_message, 'after_step'))
-                t.after_epochs.add_hook(
-                    functools.partial(log_message, 'after_epoch'))
+                for key in [EventKeys.BEFORE_EPOCH,
+                            EventKeys.BEFORE_STEP,
+                            EventKeys.AFTER_STEP_ANNEAL,
+                            EventKeys.AFTER_STEP_EVAL,
+                            EventKeys.AFTER_STEP_LOG,
+                            EventKeys.AFTER_STEP,
+                            EventKeys.AFTER_EPOCH_ANNEAL,
+                            EventKeys.AFTER_EPOCH_EVAL,
+                            EventKeys.AFTER_EPOCH_LOG,
+                            EventKeys.AFTER_EPOCH]:
+                    t.events.on(key, functools.partial(log_event, key))
 
                 t.run()
                 self.assertEqual(4, len(t._run_step.call_args_list))
@@ -156,11 +181,30 @@ class BaseTrainerTestCase(tf.test.TestCase):
                         call_payload[1][0]
                     )
 
-                self.assertEqual(
-                    ['before_epoch', 'before_step', 'after_step',
-                     'before_step', 'after_step', 'after_epoch'] * 2,
-                    logged_messages
+                expected_logged_events = sum(
+                    [
+                        [
+                            (EventKeys.BEFORE_EPOCH, epoch + 1),
+                        ] + sum([
+                            [
+                                (EventKeys.BEFORE_STEP, epoch * 2 + step + 1),
+                                (EventKeys.AFTER_STEP_EVAL, epoch * 2 + step + 1),
+                                (EventKeys.AFTER_STEP_ANNEAL, epoch * 2 + step + 1),
+                                (EventKeys.AFTER_STEP_LOG, epoch * 2 + step + 1),
+                                (EventKeys.AFTER_STEP, epoch * 2 + step + 1),
+                            ]
+                            for step in [0, 1]
+                        ], []) + [
+                            (EventKeys.AFTER_EPOCH_EVAL, epoch + 1),
+                            (EventKeys.AFTER_EPOCH_ANNEAL, epoch + 1),
+                            (EventKeys.AFTER_EPOCH_LOG, epoch + 1),
+                            (EventKeys.AFTER_EPOCH, epoch + 1)
+                        ]
+                        for epoch in [0, 1]
+                    ],
+                    []
                 )
+                self.assertListEqual(logged_events, expected_logged_events)
 
             # test re-entrant error
             with TrainLoop([], max_epoch=1) as loop:
@@ -168,11 +212,11 @@ class BaseTrainerTestCase(tf.test.TestCase):
                 t._run_step = Mock(return_value=None)
                 t._iter_steps = Mock(wraps=lambda: loop.iter_steps(df))
 
-                def reentrant_error():
+                def reentrant_error(step):
                     with pytest.raises(
                             RuntimeError, match=r'`run\(\)` is not re-entrant'):
                         t.run()
                 reentrant_error = Mock(wraps=reentrant_error)
-                t.after_steps.add_hook(reentrant_error)
+                t.events.on(EventKeys.AFTER_STEP, reentrant_error)
                 t.run()
                 self.assertTrue(reentrant_error.called)