+++ tfsnippet/scaffold/__init__.py
@@ -1,12 +1,13 @@
 from .checkpoint import *
 from .early_stopping_ import *
+from .event_keys import *
 from .logging_ import *
 from .scheduled_var import *
 from .train_loop_ import *
 
 __all__ = [
     'AnnealingVariable', 'CheckpointSaver', 'DefaultMetricFormatter',
-    'EarlyStopping', 'EarlyStoppingContext', 'MetricFormatter', 'MetricLogger',
-    'ScheduledVariable', 'TrainLoop', 'TrainLoopContext', 'early_stopping',
-    'summarize_variables', 'train_loop',
+    'EarlyStopping', 'EarlyStoppingContext', 'EventKeys', 'MetricFormatter',
+    'MetricLogger', 'ScheduledVariable', 'TrainLoop', 'TrainLoopContext',
+    'early_stopping', 'summarize_variables', 'train_loop',
 ]