+++ tfsnippet/trainer/__init__.py
@@ -1,13 +1,11 @@
 from .base_trainer import *
 from .evaluator import *
 from .feed_dict import *
-from .hooks import *
 from .loss_trainer import *
 from .trainer import *
 from .validator import *
 
 __all__ = [
-    'BaseTrainer', 'Evaluator', 'HookEntry', 'HookList', 'HookPriority',
-    'LossTrainer', 'Trainer', 'Validator', 'auto_batch_weight',
-    'merge_feed_dict', 'resolve_feed_dict',
+    'BaseTrainer', 'Evaluator', 'LossTrainer', 'Trainer', 'Validator',
+    'auto_batch_weight', 'merge_feed_dict', 'resolve_feed_dict',
 ]